Init, refactor out baseProcessor
