--- scrapers/classes/processors/tests/baseProcessor.test.js
@@ -1,8 +1,0 @@
-/*
- * This file is part of Search NEU and licensed under AGPL3.
- * See the license file in the root folder for details.
- */
-
-it("should work", () => {
-  console.log("Not implemented yet");
-});
